+++ code/test/sixsq/nuvla/server/resources/common/user_utils_test.clj
@@ -27,7 +27,7 @@
         session         (-> (ltu/ring-app)
                             session
                             (content-type "application/json"))
-        session-admin   (header session authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon")]
+        session-admin   (header session authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")]
 
     (with-redefs [email-utils/extract-smtp-cfg
                                       (fn [_]