+++ code/test/sixsq/nuvla/server/resources/deployment_lifecycle_test.clj
@@ -87,7 +87,7 @@
                                  session
                                  (content-type "application/json"))
           session-admin      (header session-anon authn-info-header
-                                     (str "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon " session-id))
+                                     (str "group/nuvla-admin group/nuvla-user group/nuvla-anon " session-id))
           session-user       (header session-anon authn-info-header
                                      (str "user/jane group/nuvla-user group/nuvla-anon " session-id))
 
@@ -465,7 +465,7 @@
                                session
                                (content-type "application/json"))
           session-admin    (header session-anon authn-info-header
-                                   "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon")
+                                   "group/nuvla-admin group/nuvla-user group/nuvla-anon")
           session-user     (header session-anon authn-info-header
                                    "user/jane group/nuvla-user group/nuvla-anon")