+++ code/test/sixsq/nuvla/server/resources/evidence_record_lifecycle_test.clj
@@ -89,7 +89,7 @@
   [f]
   (let [session-admin (-> (session (ltu/ring-app))
                           (content-type "application/json")
-                          (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))]
+                          (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))]
 
     (doseq [namespace [ns1 ns2]]
       (-> session-admin
@@ -108,7 +108,7 @@
 
   (let [session-admin (-> (session (ltu/ring-app))
                           (content-type "application/json")
-                          (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))
+                          (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))
         session-user  (-> (session (ltu/ring-app))
                           (content-type "application/json")
                           (header authn-info-header "user/jane group/nuvla-user group/nuvla-anon"))
@@ -221,7 +221,7 @@
 
   (let [session-admin       (-> (session (ltu/ring-app))
                                 (content-type "application/json")
-                                (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))
+                                (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))
         session-user        (-> (session (ltu/ring-app))
                                 (content-type "application/json")
                                 (header authn-info-header "user/jane group/nuvla-user group/nuvla-anon"))
@@ -259,7 +259,7 @@
 
   (let [session-admin (-> (session (ltu/ring-app))
                           (content-type "application/json")
-                          (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))
+                          (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))
         session-user  (-> (session (ltu/ring-app))
                           (content-type "application/json")
                           (header authn-info-header "user/jane group/nuvla-user group/nuvla-anon"))]
@@ -303,7 +303,7 @@
                        (keyword (str ns1-prefix ":" attr)) "123.456"}
         session-admin (-> (session (ltu/ring-app))
                           (content-type "application/json")
-                          (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))]
+                          (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))]
 
 
     ;; create resource for testing queries
@@ -354,8 +354,8 @@
         session-admin-form    (-> (ltu/ring-app)
                                   session
                                   (content-type "application/x-www-form-urlencoded")
-                                  (header authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon"))
-        session-admin-json    (header session-anon authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon")
+                                  (header authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon"))
+        session-admin-json    (header session-anon authn-info-header "group/nuvla-admin group/nuvla-user group/nuvla-anon")
         session-user          (header session-anon authn-info-header "user/jane group/nuvla-user group/nuvla-anon")
 
         attr1                 (ltu/random-string)